Switch build links to main branch
shipwright-io/build will be switching its default branch from "master"
to "main". This renames the file links that point to branches in GitHub.

GitHub's branch renaming feature will establish a redirect for all links
that point to the "master" branch. This change ensures that we will not
rely on the redirect in the long run.